+++ src/lib/index.ts
@@ -1,5 +1,4 @@
 export * from "./Session";
 export * from "./Response";
-export * from "./Cookie";
 export * from "./requests";
-export * from "./Client";
+export * from "./tlsClient";