+++ src/utils/index.ts
@@ -1,6 +1,4 @@
 export * from "./download";
-export * from "./koffi";
 export * from "./logger";
 export * from "./request";
-export * from "./download";
 export * from "./tlsError";