+++ src/utils/request.ts
@@ -22,7 +22,7 @@ const prefixContentTypes = new Set([
 /**
  * Determines if a request should be treated as a byte request based on its headers.
  *
- * This function checks the "Content-Type" and "Content-Transfer-Encoding" headers
+ * This function checks the "Content-Type", "Content-Transfer-Encoding" and "Content-Encoding" headers
  * to determine if the request is for binary data (such as images, audio, video,
  * application binaries, etc.). If the headers indicate binary data, it returns true.
  *
@@ -30,6 +30,8 @@ const prefixContentTypes = new Set([
  * @returns {boolean} - Returns true if the request is for binary data, otherwise false.
  */
 export function isByteRequest(headers: OutgoingHttpHeaders): boolean {
+  if (headers["content-encoding"] || headers["Content-Encoding"]) return true;
+
   const contentType = headers["content-type"] || headers["Content-Type"];
   const contentTransferEncoding =
     headers["content-transfer-encoding"] ||