+++ typings/lib/Response.d.ts
@@ -1,12 +1,12 @@
-/// <reference types="node" />
 import { TlsResponse } from "../interface";
 import { IncomingHttpHeaders } from "http";
 export declare class Response {
-    private readonly response;
-    readonly ok: boolean;
-    readonly headers: IncomingHttpHeaders;
-    readonly status: number;
-    readonly url: string;
+    ok: boolean;
+    headers: IncomingHttpHeaders;
+    status: number;
+    url: string;
+    body: string;
+    cookies: Record<string, string>;
     constructor(response: TlsResponse);
     /**
      * Returns the body of the response as a string.
@@ -20,11 +20,5 @@ export declare class Response {
      * @typeparam T - The type of the JSON object.
      * @returns A promise that resolves with the body of the response as a JSON object.
      */
-    json<T>(): Promise<T>;
-    /**
-     * Returns the cookies from the response as an object with key-value pairs.
-     *
-     * @returns An object containing cookies as key-value pairs.
-     */
-    get cookies(): Record<string, string>;
+    json<T extends unknown>(): Promise<T>;
 }