+++ typings/lib/tlsClient.d.ts
@@ -1,0 +1,2 @@
+export declare function initTLS(): Promise<void>;
+export declare function destroyTLS(): Promise<void>;