+++ typings/utils/download.d.ts
@@ -1,15 +1,12 @@
-export declare class Download {
-    private file;
-    private path;
-    private issueURL;
-    constructor(file: {
-        name: string;
-        downloadName: string;
-    }, libPath: string);
-    init(): Promise<void>;
-    private formatBytes;
-    private progress;
-    private download;
-    private extract;
-    private getLatest;
+import { FileInfo } from "../interface/native";
+export declare class LibraryDownloader {
+    private static file;
+    private static path;
+    private static readonly issueURL;
+    static retrieveLibrary(file: FileInfo, libPath: string): Promise<boolean>;
+    private static formatBytes;
+    private static progress;
+    private static download;
+    private static extract;
+    private static getLatest;
 }