+++ typings/utils/koffi.d.ts
@@ -1,6 +1,0 @@
-import { IClient } from "../interface/client";
-/**
- * Downloads and loads the native library.
- * @returns {Promise<IClient>}
- */
-export declare function load(): Promise<IClient>;