+++ typings/utils/request.d.ts
@@ -1,9 +1,8 @@
-/// <reference types="node" />
 import { OutgoingHttpHeaders } from "http";
 /**
  * Determines if a request should be treated as a byte request based on its headers.
  *
- * This function checks the "Content-Type" and "Content-Transfer-Encoding" headers
+ * This function checks the "Content-Type", "Content-Transfer-Encoding" and "Content-Encoding" headers
  * to determine if the request is for binary data (such as images, audio, video,
  * application binaries, etc.). If the headers indicate binary data, it returns true.
  *